+++ src/main/java/com/xladmt/makify/payment/dto/PaymentInitRequest.java
@@ -3,10 +3,12 @@ package com.xladmt.makify.payment.dto;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class PaymentInitRequest {
     private Long challengeId;
     private Long userId;
-    private Long paymentAmount; // 결제 금액
+    private BigDecimal paymentAmount; // 결제 금액
 }
